First draft of DecentTree: 1. Added a progress reporting class (based on the old boost    progress_display class), via progress.cpp and progress.h. 2. The constructTree() function (in StartTree::BuilderInterface)    now returns bool (true for success, false for failure)    BIONJ2009Adapter::constructTree() returns bool (true for success)    (Though I haven't added any "detect an error, return false"    Logic in BioNj::create() as yet, it's a step closer)    The constructTree() functions (in distance matrix classes    in bionj2.cpp) track progress via a progress_display    (BoundingMatrix uses two progress_display instances, the other    tracks the set up of the auxiliary S and I matrices). 3. Added a compression_level parameter to gzstreambuf::open (it    defaults to 9, not 5) (Why use lower compression levels?    CPU is cheap, but disk space is a finite resource!).  See the    call to gzsetparams in gzstreambuf::open. 4. Added compressed_length (figured out for reading, in     gzstreambuf::open()) and compressed_poisition (figured out,    during reading, in gzstreambuf::underflow()).  I haven't    Bothered (yet) to keep these updated during writing to    compressed output streams.  They're accessed via    getCompressedLength() and getCompressedPosition() getters,    on gzstreambuf and gzstreambase. 5. UPGMA_Matrix::loadMatrixFromFile uses an igzstream, checks for    errors (and returns true on success, false on failure).    Similarly loadMatrix returns bool.    Overrides of loadMatrixFromFile and loadMatrix in BIONJMatrix    must also return bool rather than void. 6. Added a setZippedOutput() function to StartTree::BuilderInterface    (And added code in StartTree::Builder and BenchmarkingBuilder    to honour it, by passing it down to setZippedOutput member    functions on matrix classes; see below).    (a) ClusterTree's writeTreeToFile is now templated (it can be        passed either an stream of an ogzstream reference);    (b) It try/catches file I/O problems, so it can return a bool        (true for success, false for failure).    (c) ClusterTree's writeTreeFile now takes an additional        (zipIt) parameter (true to compress, false not to) 7. Clustering tie-breaks (when scores are equal, the cluster join with    the smallest "imbalance" between cluster sizes is chosen; see    In particular the Position class and various getImbalance() member    functions on distance matrix classes (all in bionj2.cpp). 8. MATRIX_ALIGNMENT math uses & (MATRIX_ALIGNMENT - 1) rather than    % (MATRIX_ALIGNMENT) (it's faster, and safe so long as MATRIX_ALIGNMENT    Is a power of 2) (all this is bionj2.cpp too). 9. Distance matrix algorithms need to know their own names (in the error    Path!  When reporting I/O errors et cetera!), so there's a    getAlgorithmName() member function for that (many classes, in bionj2.cpp). 10.Added draft of DecentTree.cpp (and a DecentTree target to the    CMakeLists.txt in the utils folder).  Haven't really settled on    Parameter names (currently, -t, -in, -out, and -gz), and it doesn't    Even list the parameters if you forget to pass any (yet).

+++ utils/heapsort.h
@@ -37,7 +37,7 @@
 
 template <class V, class S>
 void constructMirroredHeap ( V* valueArray
-                            , int start, int stop
+                            , ptrdiff_t start, ptrdiff_t stop
                             , S* satteliteArray) {
     //
     //Constructs a "top-less" radix 2 max heap
@@ -45,12 +45,12 @@ void constructMirroredHeap ( V* valueArray
     //the permutation of the values
     //on satteliteArray[start..stop].
     //
-    int fudge = 2 - start;
-    for (int h = start+(stop-start)/2; h >= start; --h) {
-        int i = h;
-        V   v = valueArray[i];
-        S   s = satteliteArray[i];
-        int j = i + i + fudge;
+    ptrdiff_t fudge = 2 - start;
+    for (ptrdiff_t h = start+(stop-start)/2; h >= start; --h) {
+        ptrdiff_t i = h;
+        V         v = valueArray[i];
+        S         s = satteliteArray[i];
+        ptrdiff_t j = i + i + fudge;
         while ( j < stop ) {
             if ( j + 1 < stop ) {
                 j += ( valueArray[j] < valueArray[j+1] ) ? 1 : 0;
@@ -69,18 +69,18 @@ void constructMirroredHeap ( V* valueArray
 }
 template <class V, class S>
 void extractFromMirroredHeap ( V* valueArray
-                              , int start, int stop
+                              , ptrdiff_t start, ptrdiff_t stop
                               , S* satteliteArray) {
     //
     //Extracts a value (and some sattelite information)
     //from a radix 2 max heap.
     //
-    int fudge = 2 - start;
+    ptrdiff_t fudge = 2 - start;
     for (--stop; start<=stop; --stop) {
-        int i = stop;
-        V   v = valueArray[i];
-        S   s = satteliteArray[i];
-        int j = start;
+        ptrdiff_t i = stop;
+        V         v = valueArray[i];
+        S         s = satteliteArray[i];
+        ptrdiff_t j = start;
         while ( j < stop ) {
             if ( j + 1 < stop ) {
                 j += ( valueArray[j] < valueArray[j+1] ) ? 1 : 0;
@@ -100,7 +100,7 @@ void extractFromMirroredHeap ( V* valueArray
 
 template <class V, class S>
 void mirroredHeapsort ( V* valueArray
-                       , int start, int stop
+                       , size_t start, size_t stop
                        , S* satteliteArray) {
     //
     //Sorts valueArray[start..stop] and "mirrors"
